requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
email-validator==2.0.0
qrcode==7.4.2
reportlab==4.0.7
python-dotenv==1.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
openpyxl==3.1.2
